# src/main/java/com/informedeimagenes/db/informedeimagenes_db/entity/Ecografia.java
package com.informedeimagenes.db.informedeimagenes_db.entity;

import com.informedeimagenes.db.informedeimagenes_db.entity.base.EcografiaBase;

public class Ecografia extends EcografiaBase {

	//OVERRIDE HERE YOUR CUSTOM MAPPER
	
	
}