# src/main/java/com/informedeimagenes/db/informedeimagenes_db/entity/UserType.java
package com.informedeimagenes.db.informedeimagenes_db.entity;

import com.informedeimagenes.db.informedeimagenes_db.entity.base.UserTypeBase;

public class UserType extends UserTypeBase {

	//OVERRIDE HERE YOUR CUSTOM MAPPER
	
	
}